wallet full

+++ Source/SteamHub.ApiContract/Models/User/UpdateUserRequest.cs
@@ -8,7 +8,7 @@ public class UpdateUserRequest
 
     public string Email { get; set; }
 
-    public float WalletBalance { get; set; }
+    public decimal WalletBalance { get; set; }
 
     public float PointsBalance { get; set; }
